Change all create and edit modals to work as was intended
This method is using useRef, may want to change this to use useState in the future so it is a controlled component instead. (Couldn't get it to work when I tried the first time).

Options when creating vehicle now only show available cards and tags.

Options when updating vehicles now only show the vehicles current card and tag and also the other available cards and tags.

A "None" option was added to the selectable drop downs in the create and update vehicle modals where applicable (obviously could change). Currently doesn't work due to how backend is handling connecting and disconnecting relationships